VERSION 0.2.0 Authors: itdobro <itsdobro@gmail.com>
ОбновлениÑе:
Создана минимальная архитектура проекта, функциональные модули разделены.

Добавлено:
- Форматирование префиксов в Ðвывода (в блокноте)
- Поддержка абсолютно всех дорожек медиафайлÐов

Дополнительно:
- Константы вынесены в отдельный файл
