+++ Cargo.toml
@@ -1,7 +1,9 @@
 [package]
 name = "media_sherlock"
-version = "0.1.2"
+description = "This utility collects data about media files"
+version = "0.2.0"
 edition = "2021"
+authors = ["itdobro <itsdobro@gmail.com>"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -9,4 +11,4 @@ edition = "2021"
 serde = "1.0"
 serde_json = "1.0"
 figlet-rs = "0.1.5"
-winres = "0.1.12"
+regex = { version = "1.10.2", features = [] }
