+++ src/core/constants.rs
@@ -1,0 +1,5 @@
+// Mod with PUBLIC constants
+
+pub const APP_NAME: &str = "Media Sherlock";
+pub const VERSION: &str = "0.2.0";
+pub const AUTHOR: &str = "https://github.com/Avdushin";
